+++ src/main/java/com/jit/proxiBanqueV4/metier/IConseillerMetier.java
@@ -9,9 +9,9 @@ import com.jit.proxiBanqueV4.entites.Conseiller;
 public interface IConseillerMetier {
 	public Conseiller saveConseiller(Conseiller conseiller);
 	public List<Conseiller> listeConseillers();
-	public Conseiller getConseiller(Long idConseiller);
+	public Conseiller getConseiller(int idConseiller);
 	public boolean updateConseiller(Conseiller conseiller);
-	public boolean deleteConseiller(Long idConseiller);
+	public boolean deleteConseiller(int idConseiller);
 	public int seConnecter(String emailConseiller,String password);
 	
 	
@@ -19,8 +19,9 @@ public interface IConseillerMetier {
 	public Client createClient(Client client);
 	public Client getClient(Long idClient);
 	public boolean updateClient(Client client);
-	public List<Client> listeClients(Long idConseiller);
-	public List<Client> alertDecouvert(Long idConseiller);
+	public List<Client> listeClients(int idConseiller);
+	public List<Client> alertDecouvert(int idConseiller);
+	
 	
 
 }
